# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Wedding Invite API
  version: 1.0.0
  description: API for managing wedding invitations

paths:
  /health:
    get:
      summary: Health check
      operationId: getHealth
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /invites/{id}:
    get:
      summary: Get an invite by ID
      operationId: getInvite
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Invite found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invite"
        "404":
          description: Invite not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      summary: Accept an invite
      operationId: putInvite
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteUpdate"
      responses:
        "200":
          description: Invite updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invite"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Invite not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    HealthResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string

    Invite:
      type: object
      required:
        - people
        - additionalCount
        - accepted
      properties:
        people:
          type: array
          items:
            type: string
        additionalCount:
          type: integer
        additional:
          type: array
          maxItems: 5
          items:
            type: string
            minLength: 1
            pattern: '^[\p{Cyrillic} \-]+$'
        accepted:
          type: boolean

    InviteUpdate:
      type: object
      required:
        - accepted
      properties:
        accepted:
          type: boolean
        additional:
          type: array
          maxItems: 5
          items:
            type: string
            minLength: 1
            pattern: '^[\p{Cyrillic} \-]+$'

    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string